[Fix] LOOPCARE-2016 The buttons do not match the design in the emergency screen on iPhone 12

See merge request mobidev/loopcare-frontend!597

+++ lib/features/nutrition/application/recipe/dto/update_recipe_body.dart
@@ -9,9 +9,8 @@ abstract class UpdateRecipeBody implements _$UpdateRecipeBody {
   const UpdateRecipeBody._();
 
   const factory UpdateRecipeBody({
-    required int numberOfUnits,
+    required double numberOfUnits,
   }) = _UpdateRecipeBody;
 
-  factory UpdateRecipeBody.fromJson(Map<String, dynamic> json) =>
-      _$UpdateRecipeBodyFromJson(json);
+  factory UpdateRecipeBody.fromJson(Map<String, dynamic> json) => _$UpdateRecipeBodyFromJson(json);
 }
